+++ src/lib.rs
@@ -20,11 +20,13 @@ extern crate cairo;
 extern crate ron;
 extern crate serde;
 
-pub mod v1;
-pub mod config;
-pub mod render;
-pub mod data;
-pub mod output;
+pub mod ast;
 pub mod clock;
+pub mod config;
+pub mod data;
+pub mod env;
+pub mod output;
+pub mod render;
+pub mod typechecker;
+pub mod v1;
 pub mod vm;
-pub mod bytecode;
